remove rewrite logic
