+++ next.config.js
@@ -5,44 +5,43 @@ const withTM = require('next-transpile-modules')([
   'unist-util-find-after',
 ]);
 
+const domains = require('./domains.json');
+
+// https://vercel.com/docs/environment-variables#system-environment-variables
+const env = process.env.VERCEL_ENV;
+
+const host =
+  env === 'production' ? 'docs.page' : env === 'preview' ? process.env.VERCEL_URL : 'localhost';
+
 module.exports = withTM({
   async redirects() {
+    const redirects = domains.map(([domain, repository]) => {
+      const [organization, repo] = repository.split('/');
+
+      return {
+        // TODO handle refs? -> `/${organization}/${repo}~:ref/:path*`,
+        source: `/${organization}/${repo}/:path*`,
+        has: [
+          {
+            type: 'host',
+            value: host,
+          },
+        ],
+        destination:
+          process.env.NODE_ENV === 'production'
+            ? `https://${domain}/:path*`
+            : `http://${domain}:${process.env.PORT || 3000}/:path*`,
+        permanent: true,
+      };
+    });
+
     return [
       {
         source: '/robots.txt',
         destination: '/res/robots.txt',
         permanent: true,
       },
+      ...redirects,
     ];
-  },
-  async rewrites() {
-    return {
-      beforeFiles: [
-        // Catch any `https://docs.page` or `https://customdomain.com` requests are prepend the host
-        {
-          source: '/',
-          has: [
-            {
-              type: 'host',
-              value: '(?<host>.*)',
-            },
-          ],
-          destination: '/:host',
-        },
-      ],
-      afterFiles: [
-        // Catch all other requests and prepend the host
-        {
-          source: `/:path*`,
-          has: [
-            {
-              type: 'host',
-              value: '(?<host>.*)',
-            },
-          ],
-          destination: `/:host/:path*`,
-        },
-      ],
-    };
   },
 });
