+++ src/pages/[owner]/[name]/[[...slug]].tsx
@@ -1,0 +1,180 @@
+import React from 'react';
+import { readFileSync } from 'fs';
+import { join } from 'path';
+import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType } from 'next';
+import NextHead from 'next/head';
+import NextRouter, { useRouter } from 'next/router';
+import NProgress from 'nprogress';
+import { MDXRemoteSerializeResult } from '@invertase/next-mdx-remote/dist/types';
+
+import { Hydrate } from '../../../mdx';
+
+import { ThemeStyles } from '../../../components/ThemeStyles';
+import { Layout } from '../../../components/Layout';
+import { Error, ErrorBoundary } from '../../../templates/error';
+
+import { ConfigContext } from '../../../utils/projectConfig';
+import { IRenderError, redirect, RenderError } from '../../../utils/error';
+import { Properties, SlugProperties, SlugPropertiesContext } from '../../../utils/properties';
+import {
+  getPageContent,
+  HeadingNode,
+  PageContent,
+  PageContentContext,
+} from '../../../utils/content';
+import { getPullRequestMetadata } from '../../../utils/github';
+import { CustomDomain, CustomDomainContext } from '../../../utils/domain';
+import { getHeadTags } from '../../../utils/html';
+import { mdxSerialize } from '../../../utils/mdx-serialize';
+import { Loading } from '../../../templates/Loading';
+
+NProgress.configure({ showSpinner: false });
+NextRouter.events.on('routeChangeStart', NProgress.start);
+NextRouter.events.on('routeChangeComplete', NProgress.done);
+NextRouter.events.on('routeChangeError', NProgress.done);
+
+export default function Documentation({
+  domain,
+  source,
+  properties,
+  content,
+  error,
+}: InferGetStaticPropsType<typeof getStaticProps>): JSX.Element {
+  const { isFallback } = useRouter();
+
+  if (isFallback) {
+    return <Loading />;
+  }
+
+  if (error) {
+    return <Error {...error} />;
+  }
+
+  const tags = getHeadTags(properties, content);
+
+  return (
+    <>
+      <NextHead>{tags}</NextHead>
+      <CustomDomainContext.Provider value={domain}>
+        <ConfigContext.Provider value={content.config}>
+          <SlugPropertiesContext.Provider value={properties}>
+            <PageContentContext.Provider value={content}>
+              <ThemeStyles />
+              <Layout>
+                <ErrorBoundary>
+                  <Hydrate source={source} />
+                </ErrorBoundary>
+              </Layout>
+            </PageContentContext.Provider>
+          </SlugPropertiesContext.Provider>
+        </ConfigContext.Provider>
+      </CustomDomainContext.Provider>
+    </>
+  );
+}
+
+export const getStaticPaths: GetStaticPaths = async () => {
+  // eslint-disable-next-line prefer-const
+  let paths = [];
+
+  // Since this call can be fairly large, only run it on production
+  // and let the development pages fallback each time.
+  // if (isProduction()) {
+  //   const repositories = getRepositoriesList();
+  //   paths = await getRepositoriesPaths(repositories);
+  //   console.info(`- gathered ${paths.length} static pages.`);
+  // }
+
+  return {
+    paths,
+    fallback: true,
+  };
+};
+
+type StaticProps = {
+  domain: CustomDomain;
+  properties: SlugProperties;
+  headings: HeadingNode[];
+  source?: MDXRemoteSerializeResult;
+  content?: PageContent;
+  error?: IRenderError;
+};
+
+export const getStaticProps: GetStaticProps<StaticProps> = async ctx => {
+  console.log(ctx.params);
+  const owner = (ctx.params.owner as string).toLowerCase();
+  const name = (ctx.params.name as string).toLowerCase();
+  const slug = (ctx.params.slug || []) as string[];
+
+  let source = null;
+  const headings: HeadingNode[] = [];
+  let error: RenderError = null;
+
+  // Extract the slug properties from the request.
+  const properties = new Properties([owner, name, ...slug]);
+
+  // If the ref looks like a PR, update the details to point towards
+  // the PR owner (which might be a different repo)
+  if (properties.isPullRequest()) {
+    const metadata = await getPullRequestMetadata(
+      properties.owner,
+      properties.repository,
+      parseInt(properties.ref),
+    );
+
+    // If a PR was found, update the property metadata
+    if (metadata) {
+      properties.setPullRequestMetadata(metadata);
+    }
+  }
+
+  const content = await getPageContent(properties);
+
+  if (!content) {
+    // If there is no content, the repository is not found
+    error = RenderError.repositoryNotFound(properties);
+  } else if (content.frontmatter.redirect) {
+    // Redirect the user to another page
+    return redirect(content.frontmatter.redirect, properties);
+  } else {
+    // At this point there is a repository, however there still might not be a file for the path
+
+    // If no property ref has been set, assign the base branch (usually main or master)
+    if (!properties.ref) {
+      properties.setBaseRef(content.baseBranch);
+    }
+
+    if (content.markdown) {
+      const serialization = await mdxSerialize(content);
+
+      if (serialization.error) {
+        error = RenderError.serverError(properties);
+      } else {
+        source = serialization.source;
+        content.headings = serialization.headings as unknown as HeadingNode[];
+      }
+    } else {
+      error = RenderError.pageNotFound(properties);
+    }
+  }
+
+  // Get the array of domains from the local filesystem & match a potential domain
+  const domainsPath = join(process.cwd(), 'domains.json');
+  const domains = JSON.parse(readFileSync(domainsPath, 'utf-8')) as Array<[string, string]>;
+  const domain =
+    domains.find(
+      ([, repository]) => repository === `${properties.owner}/${properties.repository}`,
+    )?.[0] || null;
+
+  return {
+    props: {
+      domain,
+      properties: properties.toObject(),
+      source,
+      headings,
+      content,
+      error: error?.toObject() ?? null,
+    },
+    revalidate: 30,
+  };
+};
