+++ src/pages/index.tsx
@@ -1,0 +1,2 @@
+import Homepage from '../templates/homepage';
+export default Homepage;
